requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
beautifulsoup4==4.12.3
    # via -r requirements/requirements.in
jinja2==3.1.4
    # via -r requirements/requirements.in
markdown==3.6
    # via -r requirements/requirements.in
markupsafe==2.1.5
    # via jinja2
python-frontmatter==1.1.0
    # via -r requirements/requirements.in
python-slugify==8.0.4
    # via -r requirements/requirements.in
pyyaml==6.0.1
    # via python-frontmatter
soupsieve==2.5
    # via beautifulsoup4
text-unidecode==1.3
    # via python-slugify
